requirements.txt
django-bootstrap-v5
django-multiselectfield
slugify